Time: 11 ms (67.11%), Space: 50.2 MB (58.17%) - LeetHub

+++ 1011-capacity-to-ship-packages-within-d-days/1011-capacity-to-ship-packages-within-d-days.java
@@ -4,14 +4,14 @@ class Solution {
         if(days > weights.length){
             return -1;
         }
-        int low = 0;
-        int high = 0;
+        long low = 0;
+        long high = 0;
         for(int i = 0; i < weights.length; i++){
             low = Math.max(low, weights[i]);
             high += weights[i];
         }
         while(low <= high){
-            int mid = low + (high -low)/2;
+            long mid = low + (high -low)/2;
 
             if(shipCapacity(weights, days, mid)){
                 high = mid -1;
@@ -19,12 +19,12 @@ class Solution {
                 low = mid +1;
             }
         }
-        return low;
+        return (int)low;
     }
     //Helping function to check conveyor transmit given days 
-    public boolean shipCapacity(int[] weights, int days, int mid){
+    public boolean shipCapacity(int[] weights, int days, long mid){
         int k = 1;
-        int packages = 0;
+        long packages = 0;
         for(int i = 0; i < weights.length; i++){
             if(packages + weights[i] <= mid){
                 packages += weights[i];
